+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Backward incompatible changes
 - Removed HTML output of error messages and debugs
 - Will now throw an exception when `mbstring.func_overload` is set
 - Moved Glyph operator `GF_` constants in separate `\Mpdf\Fonts\GlyphOperator` class
+- Availability to set custom default CSS file
 
 Fixes and code enhancements
 ----------------------------